Move CLI classes out of util into teku module.
