+++ teku/src/main/java/tech/pegasys/teku/cli/converter/LogTypeConverter.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.cli;
+package tech.pegasys.teku.cli.converter;
 
 import org.apache.logging.log4j.Level;
 import picocli.CommandLine;
