+++ teku/src/main/java/tech/pegasys/teku/cli/slashingprotection/ImportCommand.java
@@ -18,10 +18,10 @@ import java.io.IOException;
 import java.nio.file.Path;
 import org.apache.logging.log4j.util.Strings;
 import picocli.CommandLine;
+import tech.pegasys.teku.cli.converter.PicoCliVersionProvider;
 import tech.pegasys.teku.cli.options.DataOptions;
 import tech.pegasys.teku.data.SlashingProtectionImporter;
 import tech.pegasys.teku.infrastructure.logging.SubCommandLogger;
-import tech.pegasys.teku.util.cli.PicoCliVersionProvider;
 import tech.pegasys.teku.util.config.TekuConfiguration;
 
 @CommandLine.Command(
