+++ teku/src/main/java/tech/pegasys/teku/cli/subcommand/SlashingProtectionCommand.java
@@ -14,9 +14,9 @@
 package tech.pegasys.teku.cli.subcommand;
 
 import picocli.CommandLine;
+import tech.pegasys.teku.cli.converter.PicoCliVersionProvider;
 import tech.pegasys.teku.cli.slashingprotection.ExportCommand;
 import tech.pegasys.teku.cli.slashingprotection.ImportCommand;
-import tech.pegasys.teku.util.cli.PicoCliVersionProvider;
 
 @CommandLine.Command(
     name = "slashing-protection",
