+++ teku/src/test/java/tech/pegasys/teku/cli/converter/GraffitiConverterTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.cli;
+package tech.pegasys.teku.cli.converter;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
